requirements.txt
google-generativeai
textstat
python-dotenv
Pillow
requests
